Moved reputation calculation related types and func to utils folder
